+++ Modules/02-Logging.ps1
@@ -8,7 +8,7 @@ function Write-LogMessage {
     if ($logFilePath) {
         $timestamp = Get-Date -Format "yyyy-MM-dd HH:mm:ss"
         $logMessage = "$timestamp - $message"
-        Add-Content -Path $logFilePath -Value $logMessage -ErrorAction SilentlyContinue
+        Add-Content -Path $logFilePath -Value $logMessage -Encoding UTF8 -ErrorAction SilentlyContinue
     }
 }